+++ CHANGELOG.rst
@@ -6,6 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a
 Changelog <http://keepachangelog.com/en/1.0.0/>`__.
 
+[0.0.2] - 2022-MM-DD
+--------------------
+
+### Added
+- LUCIDGAN with the GAN, GAN_adult and GAN_compas examples.
+- Output-based group metrics.
+- Tests for LUCIDGAN and group metrics.
+- New section to the README on LUCIDGAN and Metrics.
+- Docs on LUCIDGAN and Metrics.
+
+
 [0.0.1] - 2022-08-26
 --------------------